Refactor TypicalPersons to TypicalEvents
